meta crate reorg: output contract
