+++ framework/meta/src/cmd/contract/output_contract/oc_config.rs
@@ -1,104 +1,6 @@
+use super::OutputContractSettings;
+use crate::cli_args::BuildArgs;
 use multiversx_sc::abi::ContractAbi;
-
-use crate::{cli_args::BuildArgs, ei::EIVersion};
-
-use super::{multi_contract_serde_stack_size::DEFAULT_STACK_SIZE, ContractAllocator};
-
-#[derive(Debug)]
-pub struct OutputContractConfig {
-    pub default_contract_config_name: String,
-    pub contracts: Vec<OutputContract>,
-}
-
-impl OutputContractConfig {
-    pub fn main_contract(&self) -> &OutputContract {
-        self.contracts
-            .iter()
-            .find(|contract| contract.main)
-            .unwrap_or_else(|| {
-                panic!(
-                    "Could not find default contract '{}' among the output contracts.",
-                    self.default_contract_config_name
-                )
-            })
-    }
-
-    pub fn main_contract_mut(&mut self) -> &mut OutputContract {
-        self.contracts
-            .iter_mut()
-            .find(|contract| contract.main)
-            .unwrap_or_else(|| {
-                panic!(
-                    "Could not find default contract '{}' among the output contracts.",
-                    self.default_contract_config_name
-                )
-            })
-    }
-
-    pub fn secondary_contracts(&self) -> impl Iterator<Item = &OutputContract> {
-        self.contracts.iter().filter(move |contract| !contract.main)
-    }
-
-    pub fn secondary_contracts_mut(&mut self) -> impl Iterator<Item = &mut OutputContract> {
-        self.contracts
-            .iter_mut()
-            .filter(move |contract| !contract.main)
-    }
-
-    pub fn get_contract_by_id(&self, contract_id: String) -> Option<&OutputContract> {
-        self.contracts
-            .iter()
-            .find(|contract| contract.contract_id == contract_id)
-    }
-
-    pub fn get_contract_by_name(&self, contract_name: String) -> Option<&OutputContract> {
-        self.contracts
-            .iter()
-            .find(|contract| contract.contract_name == contract_name)
-    }
-
-    /// Yields the contract with the given public name.
-    pub fn find_contract(&self, contract_name: &str) -> &OutputContract {
-        self.contracts
-            .iter()
-            .find(|contract| contract.contract_name == contract_name)
-            .unwrap_or_else(|| panic!("output contract {contract_name} not found"))
-    }
-}
-
-#[derive(Clone, PartialEq, Eq)]
-pub struct OutputContractSettings {
-    /// External view contracts are just readers of data from another contract.
-    pub external_view: bool,
-
-    /// Panic messages add a lot of bloat to the final bytecode,
-    /// so they should only be used for debugging purposes.
-    pub panic_message: bool,
-
-    /// Post-processing check of the VM hooks is based on this.
-    pub check_ei: Option<EIVersion>,
-
-    /// Allocator config, i.e which allocator to choose for the contract.
-    pub allocator: ContractAllocator,
-
-    pub stack_size: usize,
-
-    /// Features that are activated on the contract crate, from wasm.
-    pub features: Vec<String>,
-}
-
-impl Default for OutputContractSettings {
-    fn default() -> Self {
-        OutputContractSettings {
-            external_view: Default::default(),
-            panic_message: Default::default(),
-            check_ei: Some(EIVersion::default()),
-            allocator: Default::default(),
-            stack_size: DEFAULT_STACK_SIZE,
-            features: Default::default(),
-        }
-    }
-}
 
 /// Represents a contract created by the framework when building.
 ///
