+++ framework/meta/src/cmd/contract/output_contract/oc_settings/oc_parse_stack_size.rs
