+++ framework/meta/src/lib.rs
@@ -4,6 +4,8 @@ pub mod cli_args;
 pub mod cmd;
 pub mod ei;
 mod folder_structure;
+mod print_util;
+mod sc_file_json;
 mod tools;
 
 #[macro_use]
