+++ framework/meta/tests/multi_contract_test.rs
@@ -1,6 +1,6 @@
 use multiversx_sc::abi::{ContractAbi, EndpointAbi};
 use multiversx_sc_meta::cmd::contract::output_contract::{
-    MultiContractConfigSerde, OutputContractConfig,
+    MultiContractConfigSerde, OutputContractGlobalConfig,
 };
 
 fn get_serialized_toml() -> MultiContractConfigSerde {
@@ -91,7 +91,7 @@ fn test_output_contract_config() {
     let serde = get_serialized_toml();
     let abi = get_contract_abi();
 
-    let contract_config = OutputContractConfig::load_from_config(&serde, &abi);
+    let contract_config = OutputContractGlobalConfig::load_from_config(&serde, &abi);
 
     assert_eq!(
         contract_config.default_contract_config_name,
